one class per file - + a bit of re-org
