+++ src/Scripts/Choice.ts
@@ -1,0 +1,5 @@
+export class Choice {
+    label = "";
+    url = "";
+    checked = false;
+}
