+++ src/Scripts/DeferredError.ts
@@ -1,0 +1,4 @@
+export class DeferredError {
+    error: Error = <Error>{};
+    message = "";
+}
