+++ src/Scripts/row/Header.ts
@@ -1,0 +1,8 @@
+export class Header {
+    constructor(header: string, value: string) {
+        this.header = header;
+        this.value = value;
+    }
+    header: string;
+    value: string;
+}
