+++ src/Scripts/row/match.ts
@@ -1,0 +1,3 @@
+export class match {
+    fieldName = ""; iToken = 0;
+}
