+++ src/Scripts/table/Other.ts
@@ -1,20 +1,6 @@
 ﻿import { iTable } from "./itable";
-import { strings } from "./Strings";
-import { Header } from "./Headers";
-import { Row } from "./Summary";
-
-export class OtherRow extends Row {
-    constructor(number: number, header: string, value: any) {
-        super(header, "", "");
-        this.number = number;
-        this.value = value;
-        this.url = strings.mapHeaderToURL(header);
-    }
-
-    [index: string]: any;
-    number: number;
-    override toString() { return this.header + ": " + this.value; }
-}
+import { Header } from "../row/Header";
+import { OtherRow } from "../row/OtherRow";
 
 export class Other extends iTable {
     private _otherRows: OtherRow[] = [];
