+++ src/Scripts/table/column.ts
@@ -1,0 +1,10 @@
+export class column {
+    constructor(id: string, label: string, columnClass: string) {
+        this.id = id;
+        this.label = label;
+        this.class = columnClass;
+    }
+    id: string;
+    label: string;
+    class: string;
+}
